% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hoop.R
\name{Hoop}
\alias{Hoop}
\title{Compute MoI of a hoop}
\usage{
Hoop(m, r, axiss)
}
\arguments{
\item{m}{mass of the object in kilogram.}

\item{r}{radius of object in meter.}

\item{axiss}{The reference axis on which rotation takes place. Options are \verb{geometrical axis},\verb{perp. to surface and center}.}
}
\value{
I moment of inertia in kg*m^2.
}
\description{
Compute MoI of a hoop
}
\examples{
Hoop(7, 2, "geometrical axis")

}
