% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoI.R
\docType{package}
\name{MoI}
\alias{MoI}
\title{MoI: A package for computing the moment of inertia for various 3D objects}
\description{
The \code{MoI} package provides several functions for calculating moment of inertia of 3D objects.
}
\section{\code{MoI} functions}{

The \code{Hoop()} function computes the moment of inertia of a hoop.
The \code{RectPlate()} function computes the moment of inertia of a rectangular plate.
The \code{Rod()} function computes the moment of inertia of a rod.
The \code{SolidCylinder()} function computes the moment of inertia of a solid cylinder.
The \code{Sphere()} function computes the moment of inertia of a sphere.
}

