% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RectPlate.R
\name{RectPlate}
\alias{RectPlate}
\title{Compute MoI of a rectangular plate}
\usage{
RectPlate(m, l1, l2, axiss)
}
\arguments{
\item{m}{mass of the object in kilogram.}

\item{l1}{length of a side of object in meter (perp. to rotation axis).}

\item{l2}{length of another side of object in meter.}

\item{axiss}{The reference axis on which rotation takes place. Options are \verb{perp. center axis},\verb{along edge}.}
}
\value{
I moment of inertia in kg*m^2.
}
\description{
Compute MoI of a rectangular plate
}
\examples{
RectPlate(7, 2, 14, "perp. center axis")

}
