% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rod.R
\name{Rod}
\alias{Rod}
\title{Compute MoI of a rod}
\usage{
Rod(m, l, axiss)
}
\arguments{
\item{m}{mass of the object in kilogram.}

\item{l}{length of object in meter.}

\item{axiss}{The reference axis on which rotation takes place. Options are \verb{perp. center},\verb{diameter of surface}.}
}
\value{
I moment of inertia in kg*m^2.
}
\description{
Compute MoI of a rod
}
\examples{
Rod(5, 7, "diameter of surface")

}
