% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolidCylinder.R
\name{SolidCylinder}
\alias{SolidCylinder}
\title{Compute MoI of a solid cylinder}
\usage{
SolidCylinder(m, r, l, axiss)
}
\arguments{
\item{m}{mass of the object in kilogram.}

\item{r}{radius of object in meter.}

\item{l}{length of object in meter.}

\item{axiss}{The reference axis on which rotation takes place. Options are \verb{geometrical axis},\verb{along surface},\verb{perp. to surface and center},\verb{diameter of surface}.}
}
\value{
I moment of inertia in kg*m^2.
}
\description{
Compute MoI of a solid cylinder
}
\examples{
SolidCylinder(5, 3, 7, "geometrical axis")

}
