% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sphere.R
\name{Sphere}
\alias{Sphere}
\title{Compute MoI of a sphere}
\usage{
Sphere(m, r, type, axiss)
}
\arguments{
\item{m}{mass of the object in kilogram.}

\item{r}{radius of object in meter.}

\item{type}{whether sphere is either \code{solid} or \code{shell}.}

\item{axiss}{The reference axis on which rotation takes place. Options are \code{diameter},\code{tangent}.}
}
\value{
I moment of inertia in kg*m^2.
}
\description{
Compute MoI of a sphere
}
\examples{
Sphere(2, 6, "solid", "tangent")

}
